Crash reporter no email given to report my faults
When I install a tweak , cydia supposed to say restart device and the app crashes. Or when I install a tweak and the phone restarts the phone crashes. So when I click on the Yalu app it says already jail broken and sometimes when the phone crashes none of the apps work. So I have to manually shut the phone down and restart. When I was listening to some music for example the phone crashed and rebooted itself. I've installed Crash Reporter and when I want to report the issue the app doesn't generate an email address that I need in order to send the Crash info.